+++ mitosheet/mitosheet/ai/recon.py
@@ -5,6 +5,7 @@ from typing import Any, Dict, List, Optional, Tuple
 
 import pandas as pd
 from pandas.testing import assert_frame_equal
+from mitosheet.ai.ai_utils import get_code_string_from_last_expression
 
 from mitosheet.errors import make_exec_error
 from mitosheet.state import DATAFRAME_SOURCE_AI, State
@@ -21,20 +22,6 @@ def is_df_changed(old: pd.DataFrame, new: pd.DataFrame) -> bool:
         return False
     except AssertionError:
         return True
-
-def get_code_string_from_last_expression(code: str, last_expression: ast.stmt) -> str:
-    code_lines = code.splitlines()
-    # NOTE; these are 1-indexed, and we need make sure we add one if they are the same, so that 
-    # we can actually get the line with our slice. Also, on earlier versions of Python, the end_lineno is
-    # not defined; thus, we must access it through the attribute getter
-    lineno = last_expression.lineno - 1
-    end_lineno = last_expression.__dict__.get('end_lineno', None)
-    if end_lineno is not None:
-        end_lineno -= 1
-        if end_lineno == lineno:
-            end_lineno += 1
-    relevant_lines = code_lines[lineno:end_lineno] 
-    return "\n".join(relevant_lines)
 
 def exec_for_recon(code: str, original_df_map: Dict[str, pd.DataFrame]) -> DataframeReconData:
     """
